requirements.txt
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.17.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1

alembic==1.6.5
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
Mako==1.1.4
psycopg2-binary==2.9.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.19
